JWT DB
Logowanie poprzez użytkowników z bazy danych, JWT, BCrypt

+++ harmony-schedule/src/main/java/pl/entpoint/harmony/authentication/JwtUserDetails.java
@@ -14,12 +14,12 @@ public class JwtUserDetails implements UserDetails {
 
   private static final long serialVersionUID = 5155720064139820502L;
 
-  private final Long id;
+  private final int id;
   private final String username;
   private final String password;
   private final Collection<? extends GrantedAuthority> authorities;
 
-  public JwtUserDetails(Long id, String username, String password, String role) {
+  public JwtUserDetails(int id, String username, String password, String role) {
     this.id = id;
     this.username = username;
     this.password = password;
@@ -31,7 +31,7 @@ public class JwtUserDetails implements UserDetails {
   }
 
   @JsonIgnore
-  public Long getId() {
+  public int getId() {
     return id;
   }
 
@@ -75,5 +75,3 @@ public class JwtUserDetails implements UserDetails {
   }
 
 }
-
-
